+++ RELEASE_NOTES.md
@@ -23,3 +23,5 @@
 * 1.2.6 - Bugfix for nullable string iput parameters.
 * 1.2.7 - Runtime SqlCommand<> introduced in SqlCommandProvider. "True" synchronous Execute implemented.
 * 1.2.8 - Downgrading back to 4.0 and FSharp.Core 4.3.0
+* 1.2.9 - Making input string into string option when allParametersOptional is suplied in SqlCommandProvider
+* 1.2.10 - Introduce runtime base class for the Record. ConcurrentDictionary for cache.